requirements.txt
apache-airflow
boto3
apache-airflow-providers-amazon
apache-airflow-providers-postgres